# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_interview_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/interview-worksheets": {
            "get": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "ddcffef2a1953bf0abbfc36adf67a592",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "4709c0b056c97f5ac9bf32096bf7b720",
                "requestBody": {
                    "request": "Create Interview worksheets",
                    "description": "Create Interview worksheets Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "1",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "company_id": {
                                        "description": "1",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "description": "Текст",
                                        "type": "string",
                                        "example": "Тест"
                                    },
                                    "form_questions": {
                                        "description": "2022-04-01",
                                        "type": "json",
                                        "example": ""
                                    },
                                    "form_answers": {
                                        "description": "2022-04-01",
                                        "type": "json",
                                        "example": ""
                                    },
                                    "status": {
                                        "description": "2022-04-01",
                                        "type": "string",
                                        "example": "new"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "description": "1",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "company_id": {
                                            "description": "1",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Текст",
                                            "type": "string",
                                            "example": "Тест"
                                        },
                                        "form_questions": {
                                            "description": "2022-04-01",
                                            "type": "json",
                                            "example": ""
                                        },
                                        "form_answers": {
                                            "description": "2022-04-01",
                                            "type": "json",
                                            "example": ""
                                        },
                                        "status": {
                                            "description": "2022-04-01",
                                            "type": "string",
                                            "example": "new"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/interview-worksheets/get-by-name/{name}": {
            "get": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "1f8054eb2e145347838bdc682ce07dfe",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Имя задания",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/interview-worksheets/get-by-userId/{user_id}": {
            "get": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "024bb340b2543a7eadc82fdb00b9c74b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/interview-worksheets/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "9f66cac2a1b977063b38868993d6543e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/interview-worksheets/{id}": {
            "delete": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "c0209854faf2b2b576e32ff104d6480b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Interview worksheets"
                ],
                "operationId": "8c4f5ddd102b741b1ba71e6d4af98c4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update Interview worksheets",
                    "description": "Update Interview worksheets Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "1",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "company_id": {
                                        "description": "1",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "description": "Текст",
                                        "type": "string",
                                        "example": "Тест"
                                    },
                                    "form_questions": {
                                        "description": "2022-04-01",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "form_answers": {
                                        "description": "2022-04-01",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "status": {
                                        "description": "2022-04-01",
                                        "type": "string",
                                        "example": "new"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "description": "1",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "company_id": {
                                            "description": "1",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Текст",
                                            "type": "string",
                                            "example": "Тест"
                                        },
                                        "form_questions": {
                                            "description": "2022-04-01",
                                            "type": "json",
                                            "example": ""
                                        },
                                        "form_answers": {
                                            "description": "2022-04-01",
                                            "type": "json",
                                            "example": ""
                                        },
                                        "status": {
                                            "description": "2022-04-01",
                                            "type": "string",
                                            "example": "new"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                },
                                                "field2": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field2 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}